requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Mail
flask_apscheduler
APScheduler
psycopg2-binary
gunicorn
google-generativeai
mailjet_rest==1.3.4